[Refactor] : domain schema 분리
Git : #33
