+++ eb_fast_api/domain/auth/register/sources/register_routers.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, HTTPException, Depends
 from eb_fast_api.domain.auth.register.sources import register_feature
 from eb_fast_api.database.sources.database import EBDataBase
-from eb_fast_api.domain.schema.sources.schema import RegisterInfo
+from eb_fast_api.domain.schema.sources.schemas import RegisterInfo
 
 
 router = APIRouter(prefix="/auth/register")
