+++ eb_fast_api/domain/home/sources/home_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from typing import List
 from eb_fast_api.domain.home.sources import home_feature
 from eb_fast_api.domain.home.sources.home_schema import ScheduleInfoList
-from eb_fast_api.domain.schema.sources.schema import ScheduleInfo
+from eb_fast_api.domain.schema.sources.schemas import ScheduleInfo
 from eb_fast_api.domain.token.sources.token_feature import getUserEmail
 from eb_fast_api.database.sources.database import EBDataBase
