+++ eb_fast_api/domain/home/sources/home_schema.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 from typing import List
-from eb_fast_api.domain.schema.sources.schema import ScheduleInfo
+from eb_fast_api.domain.schema.sources.schemas import ScheduleInfo
 
 
 class ScheduleInfoList(BaseModel):
