+++ eb_fast_api/domain/map/route/tests/route_router_test.py
@@ -1,6 +1,6 @@
 import httpx
 from eb_fast_api.domain.map.route.testings import mock_json
-from eb_fast_api.domain.map.route.sources.route_schema import RouteInfo
+from eb_fast_api.domain.schema.sources.schemas import RouteInfo
 from eb_fast_api.domain.map.route.sources import route_feature
