+++ eb_fast_api/domain/schedule/sources/schedule_feature.py
@@ -1,4 +1,4 @@
-from eb_fast_api.domain.schema.sources.schema import ScheduleInfo
+from eb_fast_api.domain.schema.sources.schemas import ScheduleInfo
 from eb_fast_api.database.sources.crud.cruds import ScheduleCRUD, PlaceCRUD
