+++ eb_fast_api/domain/schedule/sources/schedule_router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException, Depends
 from eb_fast_api.domain.schedule.sources import schedule_feature
-from eb_fast_api.domain.schema.sources.schema import ScheduleInfo
+from eb_fast_api.domain.schema.sources.schemas import ScheduleInfo
 from eb_fast_api.domain.token.sources.token_feature import getUserEmail
 from eb_fast_api.database.sources.database import EBDataBase
