+++ eb_fast_api/domain/schema/sources/login_info.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class LoginInfo(BaseModel):
+    email: str
+    password: str
