+++ eb_fast_api/domain/schema/sources/schemas.py
@@ -1,0 +1,7 @@
+from eb_fast_api.domain.schema.sources.coordi_info import CoordiInfo
+from eb_fast_api.domain.schema.sources.login_info import LoginInfo
+from eb_fast_api.domain.schema.sources.place_info import PlaceInfo
+from eb_fast_api.domain.schema.sources.register_info import RegisterInfo
+from eb_fast_api.domain.schema.sources.schedule_info import ScheduleInfo
+from eb_fast_api.domain.schema.sources.token_info import TokenInfo
+from eb_fast_api.domain.schema.sources.route_info import (StationInfo, TransportInfo, SubPathInfo, PathInfo, RouteInfo,)
