+++ eb_fast_api/domain/schema/sources/token_info.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class TokenInfo(BaseModel):
+    accessToken: str
+    refreshToken: str
